Initial v2 refactor This makes a lot of progress towards v2 with settings simplification/consolidation, error handler simplification, elimination of some static resources (more to go) and moving of primary bits into a .Shared library (in prep for a StackExchange.Exceptional.AspNetCore lib).
Note that currently, settings is in "ExceptionalSettings" and old settings are still "Settings", to maintain existing web.config compatability. It'd be nice to consolidate, but it's far from trivial to do so.

Next up: moving static resources into .Shared for embedded resourcing and starting on a netstandard2.0 build of .Shared in prep for .AspNetCore.