Add Error.LastLogDate
Rleated to #85, this adds the last date an exception occured. While CreationDate is needed for de-dupe rollups, we want to show the actual last time it happened in the dashboard...this does that.

Schema scripts for SQL and MySQL also made to be re-runnable and update schemas to v2 (more changes incoming).

INFORMATION_SCHEMA is used in SQL to make it more portable for more platforms later.