ExceptionalSettings renamed to Settings
This needed to be done for sanity, but it does break existing Web.config sections. But since the error is immediate and simple, it's a worthwhile tradeoff for a major release. Users just need to change .Settings to .ConfigSettings in their web.config. I'll add a note with the other upgrade notes on breaking changes in V2 (like SQL schemas).