ASP.NET Core: alternative to .UseDeveloperExceptionPage()
This allows a user to set up an inline throw page for local debugging and a familiar/consistent experience. Usage is:

app.UseExceptional(settings =>
{
    settings.UseExceptionalPageOnThrow = env.IsDevelopment();
});
(or just true/false, whatever conditions...)

Note that highlight.js doesn't work on this page as I'm not sure of a *reasonable* way to do that (embedding *all* of the JavaScript isn't reasonable, though embedding *only* highlight.js might be...).