Settings overhaul for de-static and ASP.NET Core
This is for per-platform settings a la MiniProfiler split with a common base. We still ultimately need a static backup here for no-context logging but the rest can be sane. Still need to handle deserialization of ASP.NET Core settings from the config properly, we can maybe do this with a custom deserializer now and tidy up a bit for the IOptions/IConfigureOptions world.