Add intitial tests
We want to test advanced binding (and ensure it keeps working), good a time as any to start the test suite.

However, this is broken due to SqlClient because who the fuck knows runtime BS. This is one of those very simple scenarios that should JUST WORK, and it still doesn't in 2.0, on the latest packages.