Fixing tests
This should have worked.
Why do I need to specify a RuntimeIdentifier in the .csproj? This causes issues for anyone running the tests on anything not-my-platform. It knows which platform I'm running on *right now*.

Throwing the error is maddening, and the error message doesn't help the user AT ALL. The runtime magical shenanigans are one of the most frustrating parts of .NET Core. For them to be such with critical libs like SqlClient is crazy at this point.