Url setter + Settings for legacy compat, e.g. from Dapper
We can increase compat with 1.x so...why not?

ConfigSettings also doesn't conflict, so let's revert that back to just Settings (but still internal) so existing web.config needs zero changes to work. This is a tadeoff in usability *inside* StackExchange.Exceptional's ASP lib (with confusion) but minimal and worth it I think.

Alsp bumping to alpha3 since this has API changes before the next NuGet release.