Move Exceptional class into ASP.NET and Core libs
This moves the Exceptional static helper up unto the per-platform libs leaving only the static settings (internally) in the base lib. This is a big change for an RC but the semantcis on load racing found in production necessitate a change. The big win here is that Exceptional.Configure and Exceptional.Settings are now per-platform strongly typed and settings that are platform specific (e.g. new features in ASP.NET Core) will be easier to add and directly expose in .Configure().