Move to ExceptionStoreSettings.TableName
This gets weird of some one-off overloads and resolves #118. Table names are now set on the same settings since this is consistent across database-based providers (as is ConnectionString). TL;DR: Normalize how this works and make it more flexible at the same time (e.g. via config binding).

This will break anyone using the old overloads on RC1 but it's best to remove these now.