Drop net451 support and move .Bind() for config down to .Shared
This resolves #117 but more important cuts off net451 support now. It's simply not reasonable going forward (as it isn't covered by netstandard2.0) and it's best to do that before this major release. I know it sucks for some people on old systems but the reality is Microsoft doesn't support it either and we're dropping support for it across the board in all libraries due to "can't use netstandard2.0 only refs".

On the .Bind() front, this adds support for confguration binding to non-ASP.NET Core scenarios, so yay for that.