Settings: ensure proper usage of statics settings for default stores.
A mismatch of static initializers ultimately resulted in Statics.Settings and Exceptional.Settings to mismatch when configuring via the lambda. This resolves that, simplifies, and unifies the two library implementations to match.

Unit tests also added to identify the non-context mismatch case. I may DRY this up a bit more (but don't want to go crazy exposing on Statics), but let's get a known working setup in first.