Initial commit
This adds everything to query filtering except stripping the querystring from the server variables (both QueryString and Full URL variables). Need to figure out the cleanest way to do this...all the helpers would be a dependency in .NET Core but maybe a regex or such is good enough here.