Make handlers run for nested exceptions (fixes #164)
Also adds tests for handlers. The issue was in the helper method only testing the top level exception in the delegate...now we properly handle nested exceptions. This shoul work for the vast majority of cases, but it is a guess/cheat and the best we can do without breaking changes.